moved test and changed alias for adldap
